refactor: split app.rs into app/{slash,chat,summary}; dedup agent complete() tail
HIGH-1: convert the oversized app.rs (~2813 production lines) into a module
directory. Extract handle_slash_command into app/slash.rs, the chat-backend
construction/detection/persistence group (build_chat_agent, detect_local_chat,
persist_chat_endpoint, config_with_chat, fetch_first_model, set_chat_config)
into app/chat.rs, and the pure display utilities (summarize_slash_tool,
summarize_json_value, scalar_or_shape, parse_plan_result, SUMMARY_MAX_FIELDS)
into app/summary.rs. AppState struct, core constructor, and the event loop stay
in app/mod.rs (now ~2122 production lines). Pure code movement, no behavior
change; a file->dir module move keeps crate::app::* paths valid.

HIGH-2: dedup the 6-line complete() tail copy-pasted across RigAgentClient,
ChatGptAgentClient, and AnthropicAgentClient into one generic helper
finish_agent_request<F, E>, generic over IntoFuture<Output = Result<String, E>>
to bridge rig's per-backend typestate. Also add const MAX_AGENT_TOKENS = 1024
(u64, to match rig's AgentBuilder::max_tokens) used by all three .max_tokens()
calls (rust-reviewer LOW).

Behavior identical; 1708 tests pass single-threaded; clippy -D warnings + fmt
--check clean; agent.rs remains the sole rig crate namer.
